websocket_HW/requirements.txt
eventlet==0.33.3
python-socketio==5.10.0
python-dotenv==1.0.0